Bring in merchant to the chain of imports

+++ apps/mappings/helpers.py
@@ -16,7 +16,7 @@ def schedule_or_delete_auto_mapping_tasks(configuration: Configuration):
     :param configuration: Workspace Configuration Instance
     :return: None
     """
-    schedule_or_delete_categories_projects_tasks(configuration)
+    schedule_or_delete_fyle_import_tasks(configuration)
     schedule_auto_map_employees(
         employee_mapping_preference=configuration.auto_map_employees, workspace_id=int(configuration.workspace_id))
     schedule_tax_groups_creation(
@@ -45,16 +45,16 @@ def validate_and_trigger_auto_map_employees(workspace_id: int):
     chain.run()
 
 
-def schedule_or_delete_categories_projects_tasks(configuration: Configuration):
+def schedule_or_delete_fyle_import_tasks(configuration: Configuration):
     """
     :param configuration: Workspace Configuration Instance
     :return: None
     """
     project_mapping = MappingSetting.objects.filter(source_field='PROJECT', workspace_id=configuration.workspace_id).first()
-    if configuration.import_categories or (project_mapping and project_mapping.import_to_fyle):
+    if configuration.import_categories or (project_mapping and project_mapping.import_to_fyle) or configuration.import_vendors_as_merchants:
         start_datetime = datetime.now()
         Schedule.objects.update_or_create(
-            func='apps.mappings.tasks.auto_import_categories_and_projects',
+            func='apps.mappings.tasks.auto_import_and_map_fyle_fields',
             args='{}'.format(configuration.workspace_id),
             defaults={
                 'schedule_type': Schedule.MINUTES,
@@ -62,8 +62,8 @@ def schedule_or_delete_categories_projects_tasks(configuration: Configuration):
                 'next_run': start_datetime
             }
         )
-    elif not configuration.import_categories and not (project_mapping and project_mapping.import_to_fyle):
+    elif not configuration.import_categories and not (project_mapping and project_mapping.import_to_fyle) and not configuration.import_vendors_as_merchants:
         Schedule.objects.filter(
-            func='apps.mappings.tasks.auto_import_categories_and_projects',
+            func='apps.mappings.tasks.auto_import_and_map_fyle_fields',
             args='{}'.format(configuration.workspace_id)
         ).delete()
